% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPBRIM.r
\name{bLP}
\alias{bLP}
\title{Bipartite label propagation (LP)}
\usage{
bLP(x, as.adjacency = TRUE)
}
\arguments{
\item{x}{Bipartite network}

\item{as.adjacency}{TRUE if x is a matrix, FALSE if it its an object with three columns}
}
\description{
Performs LP
}

