% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findModules.r
\name{findModules}
\alias{findModules}
\title{Find modules}
\usage{
findModules(M, iter = 50, sparse = TRUE, ...)
}
\description{
This function takes a matrix, and performs a search for the best
partition in modules.
}

