% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.r
\name{mostFrequent}
\alias{mostFrequent}
\title{Most frequent value}
\usage{
mostFrequent(vec, w = NA)
}
\arguments{
\item{vec}{The vector of labels}

\item{w}{The weights of each elements (currently not used)}
}
\description{
Returns the most frequent value in a vector
}

