% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.r
\name{netToAdjacency}
\alias{netToAdjacency}
\title{Edge list to adjacency matrix}
\usage{
netToAdjacency(net)
}
\arguments{
\item{net}{Three-column matrix}
}
\description{
From a three column matrix (from, to, value) to an adjacency matrix
}

