% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.r
\name{plotMatrixModules}
\alias{plotMatrixModules}
\title{Plot the modules as a matrix}
\usage{
plotMatrixModules(mod, mode = "blocks")
}
\arguments{
\item{mod}{The output of findModules}

\item{mode}{The type of plot, either blocks (different modules are colored), frames (modules are outlined), or both}
}
\description{
Plot the modules as a network
}

