% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.r
\name{spread}
\alias{spread}
\title{Spread}
\usage{
spread(v, m = 0, M = 1)
}
\arguments{
\item{v}{Values to transform}

\item{m}{New min}

\item{M}{New max}
}
\description{
Transforms values in v from m to M
}

