% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getmodules.r
\name{getmodules}
\alias{getmodules}
\title{Get modules}
\usage{
getmodules(mod)
}
\arguments{
\item{mod}{The output of findModules}
}
\description{
This function takes a mod object, and returns a list of
matrices, each corresponding to a single module
}

